+++ Minimum.py
@@ -1,0 +1,2 @@
+def find_minimum(lst):
+    return min(lst)
